cleanup and node instance fix
